Add User class
The User class defines a user with the info returned by the public
API and also adds a list of followers and followings. There are
also methods for the representation of the user and to retrieve
the contact details.

Some private API methods that require authentication are added to
get the user data and a list of blocked users, but commented,
since it only works for a single user.

Also, the APIs used for the methods that return a list come from
GET methods that are paginated with a default value of 50. This
should be covered in the future.