Add Show class
The Show class defines a show with the info returned by the
public API GET method.

The requests for shows in the API is very limited for public
access, so I only retrieve the basic data and implement a
minimal representation for the class.

There is a GET method to retrieve the shows from a specified
user, but this would be better implemented in the User class.