Add get_shows and get_favorites to User class
These new methods come from the shows API, but they are better off
in the User class, so they can be used to get more user info when
creating the instance.

Note that get_favorites is commented, since the method requires the
authentication of the user, so it is not very useful for general
requests.

Also, note that the API GET method used to implement get_favorites
is paginated and will only return the first 50 results. This
should be addressed in the future to get the complete list.